Refactor and remove deprecation on logging
Refactor _log function to make it so that this can be used for all
logging required and update to using this new method thought. Also
change WARN to WARNING and ERROR to ERR as they have been
deprecated and fix some small typos. Additionally also make the
script match packaging version.