Initialise _check_hddio before first script check
Initialise _check_hddio before first script check which remove the
case were _check_hddio initialisation incorrect states that this
is the first shutdown cycle as there are no IO reading. Also add
a small sleep to fake mode finalisation so that epoch is correctly
set-up if other check are disabled.