Refactor and enhance _ping_range
Refactor _ping_range to allow a richer RANGE set-up using real IP addresses
and ranges based of network interfaces. Note: this dose mean that the Network
interface is pinged directly rather that previously were the default root was
used. This allow for a more complex IP range check. This change requires a
teak to the plugin GUI in OMV and adds the ability to disable the plugin more
conveniently in plugin GUI. Update translation, but suspects this will need
local language adjustment to make it accurate. These changes are fully
backwardly compatible with older syntax with any user changes.