Fix bugs and Shellcheck issued in _check_docker_status
There were a set of bugs in _check_docker_status. Firstly it the NSOCKETNUMBERS
was a Docker host port this was not picked up correctly. Secondly '${ports#*:}'
was use to pick a single Docker port mapping and is should have been
'${ports%:*}' this was working due to bad test data being used. Also change to
using a better regex to pick up port mapping to stop shellcheck issues which is
an improvement.