Refactor _check_ul_dl_rate, stop using fixed sleep and add init
Refactor _check_ul_dl_rate which also remove the need for autoshutdown to still
have a tmp directory. The refactor also remove the fixed sleep time as this is
incorrect as the time for last run is depended on the time the check before
_check_ul_dl_rate take to run (these can be lengthy dependent on options
selected). Initialise _check_ul_dl_rate for each network interface before first
script check which remove the case were _check_ul_dl_rate initialisation and
incorrect states that this is the first shutdown cycle as there are no ul/dl
rate at this time.