Rework _check_networkconfig and the autosutdown.service
Rework and simplify _check_networkconfig, by reworking the dependency on
autoshutdown.service to make sure all local drive are mounted and the network
is fully online before starting autshutdown script is initialised. This
requiring has leas to a partial clean-up of _check_net_status but this is just
enough to keep it running with the new check_networkconfig set-up. Also emended
docs were required.