Add disablement of _check_net_status and _check_docker_status
Make it so that _check_net_status and _check_docker_status can be
disabled. This makes the plugin interface look more consistent
and allows the socket checkers to be disabled. Note: it is
intentional not to separate docker checker and the net checker
in the GUI as is is transparent to the user.