Remove 139 and 445 form NSOCKETNUMBERS default
Remove 139 and 445 form NSOCKETNUMBERS default. Thought the
removed port will work to detect Samba connections it is better
checked using the _check_smb_status check as this will supply
more details of the current Samba connections. Note this will only
effect new installs of the plugin, old installs will be unaffected
to make this change as backwardly compatible as possible.