Allow all OS UP interfaces in _check_networkconfig()
If an interface is not in a down state i.e. the interface is up in
the OS, but may not be connected physically connected (i.e. not
patched) and has a valid IPv4 address, then it should be considered
usable by autoshutdown as it may become active at any point.
Possible stated are 'UP' we got a cattier on the network and
working, 'UNKONWN' as it state we unsure of the state, we are up at
OS level, but there may or may not be a carrier or "DOWN" we
offline, hence, we should not consider an interface in the down
state.

Fix issue: #79