Refactor, fix and extend _check_clock
Fix and refactor  _check_clock to allow config over midnight
correctly also allow it to support minutes as well as hours. This
has been done in a way to make it as backwardly compatible as
possible. Note this may require a browser cache refresh as the
plugin GUI has been updated.