Update logging OMV to split out function name and message
Update OMV logging extraction so that function name and message
are split out this make the reading of the AutoShutdown log in OMV
easier to read and align. Also make sure that forced log which are
duplicated are ignored in OMV logging. Note this logging change
will not fully apply until OMV it restarted.