Correct Shutdown command check and Docs to use systemd
Currently autdoshutdown.default and the _check_config function in
autoshutdown check for pm-X command to Suspend, Hibernate or
Suspend-Hybrid. However, the GUI set the command to systemctl
suspend, hibernate or hydrid. This change corrects the
_check_config to testing systemctl (suspend|hibernate|hybrid) is
supported correctly from the value supplied in the config. It also
updates the autoshutdown.default to match the GUI.