Add/improve new autoshutdown exit codes
Change the exit code given by the autoshutdown script on failure
to represent the different possible error states the script could
have failed for. Update the systemd set-up to take these new exit
code into account when considering if the script should be
re-stated, as in most situation a restart would be inappropriate.