Check service is enabled in autoshutdown-restart post and move
Add a check to validate that the autoshutdown service is enabled
on a restart of the service in the post event. If the check is not
added autoshutdown will start even if it was not enabled. However,
in the event of a false start the script should shutdown as it
will detect it is disabled.

Move the script from usr/lib/systemd/system-sleep/autoshutdown-restart
to lib/systemd/system-sleep/autoshutdown-restart as it dose not look
as if it executed in the 'usr' location. this shoudl fix #87