Re-work how DEBUG couples to VERBOSE mode
VERBOSE and debug follow one another in the confutation. This does
increase the coupling between the two variables and thus we can no
logger set DEBUG separately, which is correct as it is not
supported by the configuration as DEBUG is consider script
internal. This fall shot of just having the VERBOSE variable.