Rework _check_networkconfig and drop FORCE_NIC comma support
Rework _check_networkconfig to allow for the case were a
interface my not yet have reached an operating state and/or have
an IP address allocated. This is required during a resume from
suspend or hibernating where systemd no longer honer the
network-online.target as this is only valid from a clean boot.

The rework allows the configuration of a retry loop allowing
modification of the number of retry allowed (NET_IFACE_RETRY) and
time to sleep between retry (NET_IFACE_SLEEP).

This change has also show problems with FORCE_NIC. Firstly a comma
is allowed in a network interface name and hence we cannot use a
command delimited list. However, spaces are not allowed in the
interface name so we can use a space delimited list. Secondly
_check_config did validate the details of FORCE_NIC and hence a
invalid name could have been applied. Thirdly if a list of network
interfaces was supplied in FORCE_NIC the regular expression used a
filer was to loose and could allows the incorrect by
_check_networkconfig to selection an invalid interfaces i.e. the
lo could be selected incorrectly.