Fix to allows vlan with @ in names in _check_networkconfig
Fix a bug were vlan ending with '@<interface_name>' are not accepted
as in the Kernel they are only know buy there vlan name. To keep
compatibility with older version of the script the
'@<interface_name>' will be dropped from the network interface
identification throughout the script. For example: 'eth0.42@eth0'
would know as 'eth0.42'.