Add Kernel RTC wakealarm functionality
Add functionality to allow the setting of a daily RTC drive
wakealarm to all the system to wake from it shutdown state. There
are advanced setting to set the system RTC to be used (in case
there are multiple RTC available) and override functionality to
allow the override of the default alarm time based on the shot day
format (see autoshutdown.default for more information).The
wakealarm can also allow be set using the start time of UPHOURS
and is associated override.

Note for this functionality to operate there must be a RTC
installed and ACPI or it equivalent functionality offered by the
hardware being used.

As there is an interface change the GUI will need to be reloaded
before the GUI will correctly offer wakelarm functionality.