Check if the SMB client is online or offline
Check if the SMB client is online or offline by sending a ping test to
the client. For this to work all SMB client must respond to ping traffic
or the test used will incorrectly classify the client as offline. This
changes the script having to relying on the SMB lock timeout expiring
before a shutdown is triggered on the server.