Improve/fix _log function
Remove regex from _log function and simplify this so that it fails
more gracefully if a log message is set incorrectly or fails. Also
there was a issue with the regex in the way it was searching for
the priority of the massage. This change make the _log function
more robust in the case of failure.