Correct bash regex's were spaces are used
Currently an escape sequence '\ ' is sued to denote spaces in a
regex, however this is does not seem to always work (especially
if we change localizations). Fortunately the regex bash
operator '=~' supports "extended regular expressions" so we can
change '\ ' to '[[:blank:]]' (Space and tab). This is better
supported and allow better generalisation of the regex used.