Refactor supervision loop to fix and simplify
The supervision loop does not correctly output log information on
the first run of the loop, the loop also run one extra cycle
before the loop is terminated. To address the problems we modify
the logging so that fail cycles are logged as well as when all
active check pass. In addition to fixing the logging the loop has
been simplified and the extra cycle removed.