Clean-up ss command usage and refactor _check_docker_status
Clean-up ss command usage used in _check_docker_status so that it
uses it built in filtering rather than an external awk filter.
Using ss like this is a cleaner implantation and will also account
correctly for IPv4-mapped IPv6 addresses.

Also clean-up the function to better match the implementation in
_check_net_status for ease of maintenance.