Clean-up ss command usage and refactor _check_net_status
Clean-up ss command usage used in _check_net_status so that it
uses it built in filtering rather than an external awk filter.
Also remove the -n (numeric) option so that it lookup the service
that is associated to the port (see /etc/service). As the mapping
here only cover standard ports and not all port and can be
inaccurate, allow an override to the lookup service name
'NSOCKETPORTMAP' as an advanced option. Using ss like this is a
cleaner implantation and will also account correctly for
IPv4-mapped IPv6 addresses.