Rework Debug and Verbose mode output
Remove the DEBUG variable from the script and instead used the
VERBOSE variable and the priority of the message to control the
Debug output form the _log function. This allow script
simplification. However this does have a performance impact (the
script in not performance based) it ensure the debug code is
always run and hence will not fail when verbose/fake-mode is
enabled increasing reliability.