Add robustness fixed for script plugins
Allow override of default location of plugin directory:
'/etc/autoshutdown.d' using script parameter "${3}". Make sure
we match script key values correctly and remove only ['"]
characters from the beginning or end of the script input values.
Correct regex for filenames so that all regex characters work
(include begins with '^') and correct the find maxdepth and
file type match. Finally make sure that if the content of a file is
being check that we can use '-' as the first match character.