Robust iostat output parsing
Rather than using awk to parse the text output of iostat, which is
not robust to an output/software change, as was seen when
upgrading from Debian 10 to 11. We now set iostat to output it
status using JSON output and then parse this for the values we
require using jq command. This means that the current script
should now work no Debian 10 and 11 as a side effect.