Check if a client connected to a sockets is online/offline
Add a new expert setting NSOCKETSCLIENTSCHECK that check if a
client connected to a sockets is online or offline by sending
a ping test to the client. For this to work all client must
respond to ping traffic or the test used will incorrectly
classify the client as offline. This changes the script having
to relying on the client with long connection timeout that will
have to expire before a shutdown is triggered on the server.

This feature is similar to setting CHECK_SAMBA_CLIENTS which has
long timeout.