Improve IP address detection in _check_container_status
Modernise the syntax used to detect container IP addresses when
using the 'container inspect' command as it is possible for a
container to be assigned to multiple network and hence have
multiple IP addresses. Using the old 'container inspect' syntax
multiple IP addresses were not account for and could even fail to
determine any of the container addresses. This change hence also
has to updates the 'ss' command, that detects active post, so it
can take account of the multiple network and IP addresses and
aligns the 'ss' command usage with that used in _check_net_status.
This change should address the issues highlighted in #120